tenmon_consumer/requirements.txt
certifi==2019.11.28
paho-mqtt==1.5.0
pulsar-client
pipenv==2018.11.26
psycopg2-binary==2.8.4
virtualenv==16.7.9
virtualenv-clone==0.5.3
